✨ ajuste  de apagar imagem junto com o user ou bike

+++ server/api/delete/bike.ts
@@ -1,4 +1,7 @@
+import { unlink } from 'node:fs/promises'
 import { IdSchema } from '~/schemas/id'
+
+const { FILES_PATH } = useRuntimeConfig()
 
 export default defineEventHandler(async event => {
   const { user } = await requireUserSession(event, { statusCode: 401, message: 'Você não tem pemissão para executar essa ação' })
@@ -10,6 +13,15 @@ export default defineEventHandler(async event => {
   if(!body.success) throw createError({ status: 401, message: body.error.errors[0].message })
 
   const { id } = body.data
+
+  const bike = await Bikes.findById(id)
+
+  if(!bike) throw createError({ status: 401, message: 'Bicicleta não Existe' })
+
+  if(bike.foto){
+    // eslint-disable-next-line security/detect-non-literal-fs-filename
+    await unlink(`${FILES_PATH}bikes/${bike.foto}`)
+  }
 
   await Bikes.findByIdAndDelete(id)
     .catch(() => { throw createError({ status: 401, message: 'Bicicleta não Existe' }) })
